Moved HDUSD_ADDON_PT_preferences to properties/preferences.py Cleanup code. Added function get_addon_pref. Added preferences.py to CMakeLists.txt.
