+++ extern/hdusd/addon/ui/usd_nodes.py
@@ -83,7 +83,8 @@ class HDUSD_NODE_PT_usd_nodetree_dev(HdUSD_ChildPanel):
 
     @classmethod
     def poll(cls, context):
-        return context.preferences.addons['hdusd'].preferences.dev_tools
+        from ..properties.preferences import get_addon_pref
+        return get_addon_pref().dev_tools
 
     def draw(self, context):
         layout = self.layout
